Many improvements to the OSM<->GTFS linking step
Linking is now done with the creategraph utility, and uses Python Rtree to
find nearest nodes, for an order of magnitude speed up. Also, we now link
inside way boundaries, splitting the way if necessary. This fixes a number
of weird situations that could crop up where we would link long distances
to a way.